文字列を配列に変換したりする関数を kdb パッケージに移動
