+++ main.go
@@ -1,4 +1,4 @@
-package csv2sql
+package main
 
 import (
 	"database/sql"
@@ -22,6 +22,7 @@ import (
 	_ "github.com/lib/pq"
 	"github.com/pkg/errors"
 	migrate "github.com/rubenv/sql-migrate"
+	"github.com/sylms/csv2sql/kdb"
 	"golang.org/x/text/encoding/japanese"
 	"golang.org/x/text/transform"
 )
@@ -32,32 +33,6 @@ var (
 		Box: packr.New("migrations", "./migrations"),
 	}
 	now time.Time
-)
-
-const (
-	// 科目履修生申請可否
-	// ×
-	creditedAuditorsCross = iota
-	// △
-	creditedAuditorsTriangle
-	// 空
-	creditedAuditorsEmpty
-)
-
-const (
-	// 開講時期
-	_               = iota
-	termSpringACode // 春A: 1
-	termSpringBCode
-	termSpringCCode
-	termFallACode
-	termFallBCode
-	termFallCCode
-	termSummerVacationCode
-	termSpringVacationCode
-	termAllCode
-	termSpringCode
-	termFallCode
 )
 
 const (
@@ -215,34 +190,34 @@ func csvToCoursesStruct(reader io.ReadCloser) ([]Courses, error) {
 			continue
 		}
 
-		term := TermParser(row.Term)
+		term := kdb.TermParser(row.Term)
 
-		termInt, err := TermStrToInt(term)
+		termInt, err := kdb.TermStrToInt(term)
 		if err != nil {
 			return nil, err
 		}
 
-		creditedAuditors, err := creditedAuditorsParser(row.CreditedAuditors)
+		creditedAuditors, err := kdb.CreditedAuditorsParser(row.CreditedAuditors)
 		if err != nil {
 			return nil, err
 		}
 
-		csvUpdatedAt, err := csvStringDateParser(row.UpdatedAt)
+		csvUpdatedAt, err := kdb.DateParser(row.UpdatedAt)
 		if err != nil {
 			return nil, err
 		}
 
-		standardRegistrationYearParser, err := standardRegistrationYearParser(row.StandardRegistrationYear)
+		standardRegistrationYearParser, err := kdb.StandardRegistrationYearParser(row.StandardRegistrationYear)
 		if err != nil {
 			return nil, err
 		}
 
-		period, err := PeriodParser(row.Period)
+		period, err := kdb.PeriodParser(row.Period)
 		if err != nil {
 			return nil, err
 		}
 
-		instructor, err := instructorParser(row.Instructor)
+		instructor, err := kdb.InstructorParser(row.Instructor)
 		if err != nil {
 			return nil, err
 		}
@@ -293,127 +268,10 @@ func execMigrate() error {
 	return nil
 }
 
-func TermParser(termString string) []string {
-	res := []string{}
-	if termString == "" {
-		return []string{}
-	}
-	var re *regexp.Regexp
-	re = regexp.MustCompile(`(春A|春AA|春AA|春AB|春BA|春AC|春CA|春ABC)`)
-	if re.MatchString(termString) {
-		res = append(res, "春A")
-	}
-	re = regexp.MustCompile(`(春B|春BA|春AB|春BB|春BB|春BC|春CB|春ABC)`)
-	if re.MatchString(termString) {
-		res = append(res, "春B")
-	}
-	re = regexp.MustCompile(`(春C|春CA|春AC|春CB|春BC|春CC|春CC|春ABC)`)
-	if re.MatchString(termString) {
-		res = append(res, "春C")
-	}
-	re = regexp.MustCompile(`(秋A|秋AA|秋AA|秋AB|秋BA|秋AC|秋CA|秋ABC)`)
-	if re.MatchString(termString) {
-		res = append(res, "秋A")
-	}
-	re = regexp.MustCompile(`(秋B|秋BA|秋AB|秋BB|秋BB|秋BC|秋CB|秋ABC)`)
-	if re.MatchString(termString) {
-		res = append(res, "秋B")
-	}
-	re = regexp.MustCompile(`(秋C|秋CA|秋AC|秋CB|秋BC|秋CC|秋CC|秋ABC)`)
-	if re.MatchString(termString) {
-		res = append(res, "秋C")
-	}
-	re = regexp.MustCompile(`(夏季休業中)`)
-	if re.MatchString(termString) {
-		res = append(res, "夏季休業中")
-	}
-	re = regexp.MustCompile(`(春季休業中)`)
-	if re.MatchString(termString) {
-		res = append(res, "春季休業中")
-	}
-	re = regexp.MustCompile(`(通年)`)
-	if re.MatchString(termString) {
-		res = append(res, "通年")
-	}
-	re = regexp.MustCompile(`(春学期)`)
-	if re.MatchString(termString) {
-		res = append(res, "春学期")
-	}
-	re = regexp.MustCompile(`(秋学期)`)
-	if re.MatchString(termString) {
-		res = append(res, "秋学期")
-	}
-	return res
-}
-
-// カンマ区切りで担当教員を配列で返す
-func instructorParser(instructors string) ([]string, error) {
-	res := strings.Split(instructors, ",")
-	return res, nil
-}
-
-func creditedAuditorsParser(CreditedAuditors string) (int, error) {
-	if CreditedAuditors == "×" {
-		return creditedAuditorsCross, nil
-	} else if CreditedAuditors == "△" {
-		return creditedAuditorsTriangle, nil
-	} else if CreditedAuditors == "" {
-		return creditedAuditorsEmpty, nil
-	} else {
-		return -1, errors.New("invalid input:CreditedAuditors input")
-	}
-}
-
-// KdB からエクスポートした CSV に含まれている更新日時カラムのものを time.Time に変換する
-func csvStringDateParser(date string) (time.Time, error) {
-	const layout = "2006-01-02 15:04:05"
-	jst, _ := time.LoadLocation("Asia/Tokyo")
-	t, err := time.ParseInLocation(layout, date, jst)
-	if err != nil {
-		return time.Time{}, errors.WithStack(err)
-	}
-	return t, nil
-}
-
 func getDateTimeNow() time.Time {
 	jst, _ := time.LoadLocation("Asia/Tokyo")
 	now := time.Now().In(jst)
 	return now
-}
-
-// 開講時期を数値に変換
-// 別テーブルなどで管理するのが適切（？）
-func TermStrToInt(term []string) ([]int, error) {
-	res := []int{}
-	for _, t := range term {
-		switch t {
-		case "春A":
-			res = append(res, termSpringACode)
-		case "春B":
-			res = append(res, termSpringBCode)
-		case "春C":
-			res = append(res, termSpringCCode)
-		case "秋A":
-			res = append(res, termFallACode)
-		case "秋B":
-			res = append(res, termFallBCode)
-		case "秋C":
-			res = append(res, termFallCCode)
-		case "夏季休業中":
-			res = append(res, termSummerVacationCode)
-		case "春季休業中":
-			res = append(res, termSpringVacationCode)
-		case "通年":
-			res = append(res, termAllCode)
-		case "春学期":
-			res = append(res, termSpringCode)
-		case "秋学期":
-			res = append(res, termFallCode)
-		default:
-			return nil, fmt.Errorf("invalid term string: %s", t)
-		}
-	}
-	return res, nil
 }
 
 func insert(tx *sqlx.Tx, courses []Courses) error {
@@ -496,97 +354,4 @@ func insert(tx *sqlx.Tx, courses []Courses) error {
 	}
 
 	return nil
-}
-
-// 標準履修年次を配列として返す
-// 中黒をハイフンとして解釈すると"?" or "int" or "int-int" になるので，"?", "int" はそのまま
-// それ以外は間全てをいれる
-func standardRegistrationYearParser(yearString string) ([]string, error) {
-	year := []string{}
-	yearString = strings.Replace(yearString, "ー", "-", -1)
-	yearString = strings.Replace(yearString, "・", "-", -1)
-	yearString = strings.Replace(yearString, "～", "-", -1)
-	yearString = strings.Replace(yearString, "~", "-", -1)
-	yearString = strings.Replace(yearString, "、", ",", -1)
-	yearString = strings.Replace(yearString, " ", "", -1)
-
-	if len(yearString) == 1 {
-		year = append(year, yearString)
-	} else {
-		minYear, err := strconv.Atoi(string(yearString[0]))
-		if err != nil {
-			return []string{}, err
-		}
-		maxYear, _ := strconv.Atoi(string(yearString[2]))
-		if err != nil {
-			return []string{}, err
-		}
-		for i := minYear; i <= maxYear; i++ {
-			year = append(year, strconv.Itoa(i))
-		}
-	}
-
-	return year, nil
-}
-
-// 時間割をいい感じにして配列で
-func PeriodParser(periodString string) ([]string, error) {
-	period := []string{}
-	periodString = strings.Replace(periodString, " ", "", -1)
-	periodString = strings.Replace(periodString, "　", "", -1)
-	periodString = strings.Replace(periodString, "ー", "-", -1)
-	periodString = strings.Replace(periodString, "・", "", -1)
-	periodString = strings.Replace(periodString, ",", "", -1)
-	periodString = strings.Replace(periodString, "集中", "集0", -1)
-	periodString = strings.Replace(periodString, "応談", "応0", -1)
-	periodString = strings.Replace(periodString, "随時", "随0", -1)
-
-	for i := 1; i <= 8; i++ {
-		listPeriod := strconv.Itoa(i)
-		for j := i + 1; j <= 8; j++ {
-			listPeriod = listPeriod + strconv.Itoa(j)
-			spanPeriod := strconv.Itoa(i) + "-" + strconv.Itoa(j)
-			periodString = strings.Replace(periodString, spanPeriod, listPeriod, -1)
-		}
-	}
-
-	for i := 0; i <= 8; i++ {
-		for _, dayOfWeek := range []string{"月", "火", "水", "木", "金", "土", "日", "応", "随", "集"} {
-			beforeStr1 := strconv.Itoa(i) + dayOfWeek
-			beforeStr2 := dayOfWeek + strconv.Itoa(i)
-			afterStr1 := strconv.Itoa(i) + "," + dayOfWeek
-			afterStr2 := dayOfWeek + ":" + strconv.Itoa(i)
-			periodString = strings.Replace(periodString, beforeStr1, afterStr1, -1)
-			periodString = strings.Replace(periodString, beforeStr2, afterStr2, -1)
-		}
-	}
-	if len(periodString) == 0 {
-		return period, nil
-	}
-	strList := strings.Split(periodString, ",")
-
-	for _, str := range strList {
-		strList2 := strings.Split(str, ":")
-		if len(strList2) != 2 {
-			fmt.Println("-" + periodString + "-")
-			return nil, errors.New("unexpected period input : " + str)
-		} else {
-			dayOfWeek := strList2[0]
-			timeTimetable := strList2[1]
-			for i := 0; i < len([]rune(dayOfWeek)); i++ {
-				for j := 0; j < len([]rune(timeTimetable)); j++ {
-					inputStr := string([]rune(dayOfWeek)[i]) + string([]rune(timeTimetable)[j])
-					inputStr = strings.Replace(inputStr, "集0", "集", -1)
-					inputStr = strings.Replace(inputStr, "集", "集中", -1)
-					inputStr = strings.Replace(inputStr, "随0", "随", -1)
-					inputStr = strings.Replace(inputStr, "随", "随時", -1)
-					inputStr = strings.Replace(inputStr, "応0", "応", -1)
-					inputStr = strings.Replace(inputStr, "応", "応談", -1)
-					period = append(period, inputStr)
-				}
-			}
-		}
-	}
-
-	return period, nil
 }
